Refactor protocol
